+++ green-hub/src/App.jsx
@@ -15,11 +15,23 @@ import { Routes, Route } from "react-router-dom";
 import { Toaster } from "react-hot-toast";
 import Prizes from "./components/Prizes/Prizes";
 import {HeartsContext} from "./context/HeartsContext";
-import { useState} from 'react';
+import { useState,useEffect} from 'react';
+import useAuth from "./context/AuthContext";
+import { db } from "./config/firebase";
+import { doc, collection,setDoc,deleteDoc ,getDoc,query,getDocs,where} from "firebase/firestore";
 
 function App() {
+      const { currentUser } = useAuth();
     const [clickedHearts, setClickedHearts] = useState([]);
     const hearts = {clickedHearts, setClickedHearts};
+     useEffect(() => {
+   if (currentUser?.uid){
+	const q = query(collection(db, "users_hearts"), where("uid", "==", currentUser.uid));
+	getDocs(q).then(({docs}) => {
+	    setClickedHearts(docs.map(single_row => single_row.data().aid));
+	});
+      }
+  },[currentUser]);
   return (
     <>
     <HeartsContext.Provider value={hearts}>
